backend/requirements.txt
flask>=3.0
flask-socketio>=5.3
requests>=2.31
pyzmq>=25.0
